% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefit.R
\name{perturbate_knn}
\alias{perturbate_knn}
\title{Generate perturbated expression by k-NN data}
\usage{
perturbate_knn(expression, strength = 1)
}
\arguments{
\item{expression}{The original expression. The rows and columns
correspond to samples and features. The expression is normalized
count of features.}

\item{strength}{How much perturbated. \code{0.0} is weak. \code{1.0} is strong.}
}
\value{
A perturbated expression as a \code{matrix}. The matrix's
expression values are perturbated from the original expression
values. The shape of the matrix is the same as the original
expression. The dimension names of the matrix are also the same as
the original expression.
}
\description{
Generate perturbated expression from the original
expression based on k-NN (k-nearest neighbor) data.
}
\note{
This is an API for advanced users. This API may be changed.
}
