% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefit.R
\name{perturbate_poisson}
\alias{perturbate_poisson}
\title{Generate perturbated counts by the Poisson distribution}
\usage{
perturbate_poisson(counts, strength = 1)
}
\arguments{
\item{counts}{The original counts. The rows and columns correspond
to samples and features. The values are count of features.}

\item{strength}{How much perturbated. \code{0.0} is weak. \code{1.0} is strong.}
}
\value{
A perturbated counts as a \code{matrix}. The matrix's counts are
perturbated from the original counts. The shape of the matrix is
the same as the original counts. The dimension names of the
matrix are also the same as the original counts.
}
\description{
Generate perturbated counts from the original counts
by the Poisson distribution.
}
\note{
This is an API for advanced users. This API may be changed.
}
